Added a template specialization for reference types.
Added a template specialization for reference types. Still debating on whether the many specializations of the reflect class should inherit from a base class or not, considering the fact that they are designed to be final - that is, they themselves are designed not to be inherited from; however, there is a lot of repeated code.